:hammer: Extract recovery orchestration
