+++ src/SFA.DAS.Zendesk.Monitor.Function/ZenWatchDurableFunction.cs
@@ -1,5 +1,4 @@
 using FluentValidation;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
@@ -24,7 +23,6 @@ namespace ZenWatchFunction
 
     public class WatcherOrchestration
     {
-        private static readonly string WatcherInstance = "{8B2772F1-0A07-4D64-BEBE-1402520C0BD0}";
         private static readonly RetryOptions retry = new RetryOptions(TimeSpan.FromSeconds(1), 5);
 
         [FunctionName("NotifyTicket")]
@@ -44,32 +42,6 @@ namespace ZenWatchFunction
                 var instanceId = await starter.StartNewAsync(nameof(ShareListedTickets), new[] { ticket.Id });
                 return starter.CreateCheckStatusResponse(request, instanceId);
             }
-        }
-
-        [FunctionName("BackgroundTaskEntryPoint")]
-        public static Task Run(
-            [TimerTrigger("%MonitorCronSetting%")] TimerInfo timer,
-            [OrchestrationClient] DurableOrchestrationClient starter,
-            ILogger log)
-        {
-            return GetSingleInstance(starter, log);
-        }
-
-        private static async Task<DurableOrchestrationStatus> GetSingleInstance(DurableOrchestrationClient starter, ILogger log)
-        {
-            var instance = await starter.GetStatusAsync(WatcherInstance);
-
-            if (instance?.OrchestrationIsRunning() != true)
-            {
-                log.LogInformation("Starting Watcher orchestration");
-                await starter.StartNewAsync(nameof(ShareAllTickets), WatcherInstance, null);
-            }
-            else
-            {
-                log.LogWarning("Watcher orchestration is already running");
-            }
-
-            return instance;
         }
 
         [FunctionName(nameof(ShareAllTickets))]
